edge/requirements.txt
paho-mqtt==2.1.0
pillow>=10.0.0
torch>=2.2.0
torchvision>=0.17.0
